Add registration tests (#9)

+++ cypress.config.ts
@@ -2,34 +2,27 @@ import { defineConfig } from 'cypress'
 import * as fs from 'fs'
 
 export default defineConfig({
-  retries: 2,
   video: false,
   videosFolder: 'tests/cypress/videos',
   screenshotsFolder: 'tests/cypress/screenshots',
   fixturesFolder: 'tests/cypress/fixture',
   downloadsFolder: 'tests/cypress/downloads',
   e2e: {
-    setupNodeEvents(on: Cypress.PluginEvents, config: Cypress.PluginConfigOptions) {
-      on('task', {
-        activateCypressEnvFile(): null {
-          if (fs.existsSync('.env.cypress')) {
-            fs.renameSync('.env', '.env.backup')
-            fs.renameSync('.env.cypress', '.env')
-            setAppKey('.env.backup', '.env')
-          }
+    setupNodeEvents(on: Cypress.PluginEvents): void {
+      on('before:run', () => {
+        if (fs.existsSync('.env.cypress') && process.env.CYPRESS_ENV !== 'ci') {
+          fs.renameSync('.env', '.env.backup')
+          fs.renameSync('.env.cypress', '.env')
+          setAppKey('.env.backup', '.env')
+        }
+      })
 
-          return null
-        },
-
-        activateLocalEnvFile(): null {
-          if (fs.existsSync('.env.backup')) {
-            fs.renameSync('.env', '.env.cypress')
-            fs.renameSync('.env.backup', '.env')
-            setAppKey('.env.example', '.env.cypress')
-          }
-
-          return null
-        },
+      on('after:run', () => {
+        if (fs.existsSync('.env.backup') && process.env.CYPRESS_ENV !== 'ci') {
+          fs.renameSync('.env', '.env.cypress')
+          fs.renameSync('.env.backup', '.env')
+          setAppKey('.env.example', '.env.cypress')
+        }
       })
     },
     baseUrl: 'http://localhost',
